+++ plugins/item.lua
@@ -105,7 +105,20 @@ end
 
 Item.spawn_drop = function(item, x, y, target)
     local obj = Item.get_data(item).object_id
-    if obj then return gm.item_drop_object(obj, x, y, target, false) end
+    if obj then gm.item_drop_object(obj, x, y, target, false) end
+
+    -- Look for drop (because gm.item_drop_object does not actually return the instance for some reason)
+    -- The drop spawns 40 px above y parameter
+    local drop = nil
+    local drops = Instance.find_all(gm.constants.oCustomObject_pPickupItem)
+    for _, d in ipairs(drops) do
+        if math.abs(d.x - x) <= 1.0 and math.abs(d.y - (y - 40.0)) <= 1.0 then
+            drop = d
+            break
+        end
+    end
+
+    return drop
 end